refactor: 전체적인 스타일 리팩토링
