+++ package.json
@@ -21,7 +21,8 @@
     "react-intersection-observer": "^9.8.2",
     "react-redux": "^9.1.1",
     "redux": "^5.0.1",
-    "redux-logger": "^3.0.6"
+    "redux-logger": "^3.0.6",
+    "sass": "^1.75.0"
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
