+++ src/app/(route)/mydashboard/_components/SideMenu.module.css
@@ -7,6 +7,9 @@
   border-right: 0.2rem solid var(--gray_D9D9D9);
   transition: width 0.3s ease;
   white-space: nowrap;
+
+  flex-shrink: 0;
+  flex-grow: 0;
 }
 
 .logoFrame {
