+++ src/app/_components/_navbar/_dashboardNavbar/DashboardTypeNavBar.module.scss
@@ -1,15 +1,29 @@
+@function calc-navbar-width($rem) {
+  @return 10 * $rem;
+}
+
 .navbarWrapper {
-  width: 100%;
+  width: calc(100vw - #{calc-navbar-width(30)}px);
   height: 7rem;
+  top: 0;
+  left: 0;
+  box-sizing: border-box;
+
   background-color: var(--white_FFFFFF);
   display: flex;
+  position: sticky;
   justify-content: space-between;
   align-items: center;
   border-bottom: 0.1rem solid var(--gray_D9D9D9);
   padding: 0 6rem;
   overflow: hidden;
 
+  @media (max-width: 1023px) {
+    width: calc(100vw - #{calc-navbar-width(16)}px);
+  }
+
   @media (max-width: 767px) {
+    width: calc(100vw - #{calc-navbar-width(6.7)}px);
     justify-content: flex-end;
   }
 }
@@ -22,6 +36,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
+  white-space: nowrap;
 
   @media (max-width: 767px) {
     display: none;
